Add nameprep profile
Implements the stringprep profile described in RFC3491.

This implementation does not currently handle case-folding.  Case-folding
is not currently supported by the Rust standard libraries (see
https://github.com/rust-lang/rfcs/issues/791 ), and this implementation
compromises with `to_lowercase`, which works for ASCII and a few other
random cases.  A compliant implementation is TBD.

We also do not properly handle unassigned code-points.  This applies
to both the `nameprep` and `saslprep` implementation.  Fixing this is
another TBD.