Fix the fast path for nameprep
char::is_ascii_lowercase() only returns true for alphabetical characters
which are lowercase, so also add digits, '.' and '-' which are the only
characters allowed in a non-IDN domain name.

I have also added a test to not have that regress.

This was found with this merge request in the jid crate:
https://gitlab.com/xmpp-rs/xmpp-rs/-/merge_requests/205