Make this crate no_std
As almost nothing actually depends on std, there is no need to keep a
dependency on it.  This allows users such as the jid crate to actually
be no_std as well.

The only exception is std::error::Error, which is now exposed as
core::error::Error in nightly but not yet stabilized, see
https://github.com/rust-lang/rust/issues/103765